requirements.txt
requests==2.23.0
python-telegram-bot==12.6.1
youtube-dl==2020.3.24